requirements.txt
google-generativeai>=0.3.0
python-dotenv>=0.19.0
requests>=2.31.0
matplotlib
pandas
numpy
streamlit==1.31.1
google-generativeai==0.3.1
Pillow==10.0.0
